+++ docs/source/releasenotes.rst
@@ -7,8 +7,21 @@ Release notes
 `Released <https://pypi.org/project/rpaframework/#history>`_
 +++++++++
 
+11.6.1 - 12 Nov 2021
+--------------------
+
+- Library **RPA.PDF**:
+
+  - Fix non empty or junk XML dumping on PDF parsing (:issue:`287`)
+  - Fixed by ``rpaframework-pdf`` **0.10.0** (included in this release)
+
+- Library **RPA.Email.ImapSmtp**:
+
+  - Fix sender name encoding when using ``Send Message`` keyword (:issue:`279`)
+  - Fix filename encoding when using ``Save Attachment``/``Save Attachments`` keywords (:issue:`290`)
+
 11.6.0 - 4 Nov 2021
-------
+-------------------
 
 - Library **RPA.Robocorp.WorkItems**: Keyword ``Parse Work Item From Email`` for
   retrieving the input item dictionary payload from the sent e-mail JSON body which